replace more int precision clocks/sample rates with double fp ones
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@5203 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/sound/5110intf.h
@@ -22,7 +22,7 @@ READ_HANDLER( tms5110_status_r );
 int tms5110_ready_r(void);
 
 void tms5110_reset(void);
-void tms5110_set_frequency(int frequency);
+void tms5110_set_frequency(double frequency);
 
 #endif
